Squashed 'cadical/cppsrc/' changes from 14593f819..4198d817d
4198d817d removed some white space
be6256c40 one less warning level for clang on MacOS
e0bff9d93 using
0ad02bd2e added mobical processes
4f80800ab added
627b34975 Bumped year in version.
2673910ed fixed
45240a001 formatted
8b5f84bc1 added real default configuration as first
856346909 need cecho and also disabling -fsanitize=undefined by default
9392588fa fix inverted if condition
2abdea8cd Don't ignore conflicts in warmup below deciding assumption and constrain
0072bce59 Add missing set to did_external_prop
919c60d63 congruence closure: need to allow tautologies
5ea26f7ab Incorrect assertion
b263d6de4 More aggressively set did_external_prop for out-of-order
5578bf80a more
a7fd8af66 Mobical: don't force backtrack on level 0
3d4f5d04a mobical: ensure backtracking at least one level
513797801 Disconnection of UP must also reset notified
15afed551 mobical: add paranoia assertion
760e43030 mobical: only disconnect existint external propagator
669bfe9df mobical: only decide unset literals
2b349639c remove too aggressive flag under clang
3269f7114 fix clang warnings of unreachable code
618ca1b40 new CC showing
3e49a67b8 installed the configure link again
7a43b0f73 macos warning
7dae6eae2 cleaning up mock-propagator in ResetCall for mobical
ddf7acc58 skipping parcomwrite test on Darwin/Apple
6253fc1ce Improved shared library and dynamic binary built
ac18b6c3a Improved
bba797df2 Updated news.
1d5b56632 got skipping functionality
a7d55ef21 fixed
d10d1bad4 contract check warning
8104e8c99 fixed building loop
dc0bd7c6f bold
c9055deec more
15810f472 fixed
b7b6db3d3 hidden
20ac82927 fixed
6073644ab more fixes
da39f912b removed warning
067cfff0b more fixes
57b9a4065 worked on configuration script
3383d1939 parcompwrite test
7d47fb5dd Fixed missing initialization of stats.
28fe4743d fixed Apple warnings and reformatted
b72f93915 Revert "Cannot cherry pick `deduplicatedinit` (f6a0b1902)."
4f7d0837b Cannot cherry pick `deduplicatedinit` (f6a0b1902). Thus removed mentioning this new feature from NEWS.md.
e0dd388e6 more verbose error message (fixed Florian's wrong formatting of the error message)
0f335b3aa fix some warnings (#149)
6660b6837 Manually merged 'alignof' for clause arena (added by Mathias in 0d4e7af3f)
42befcf99 fix memory leak in configure script
c4b8d11f8 Manual merges and fixes to NEWS (from api_contracts)
516ebc839 extend cb_propagate comment
3285ea706 tuning some comments
b5a121ce9 typo
56eefa2d7 add description of changes to NEWS
3c60c9cad example external propagators are added
be65dca13 add external (reason) clause requirements
28f87fa6f update cb_add_reason_lit comment
aab45e895 FIX buggy copied condition in remove-observed-var (cherry picked but already fixed before)
6724e28b3 extend cb_propagate comment
2007d16f9 add cb_propagate requirement
b10c7ba9a cb_decide comment extended
8eb2167e6 add cb_decide requirements
21ce89c03 update force backtrack signature
9a067542a add more requirements
99fa8113d remove empty notification calls
0d2f2ee4c API test example propagator added
810a04651 update add/remove observed var requirements (but fixed observe IPASIR-UP bug during merge)
2a6ec880e fix remove_observed_var assertion
2e319d926 rename tainted_literal to changed_val (TODO: find a better name)
a60cd0591 merged version
cc84d26d8 starting to update versions
282831925 fix accidental reodering
b4b412b34 fix incorrect DRAT proof
4be6c5379 FIX buggy copied condition in remove-observed-var
de82453b6 went over news
b561c2db8 negate condition in the require
0edff1704 better check for properly declared variables
0e9c0a800 check fails immediatly instead of the second time
f571fcef4 Describe changes in 2.2.1
17176df13 formatted
02c5c0bd2 fix warning
4676499c8 don't delete watches in deduplication
01dcde640 fix partitioning of garbage in deduplication
d3e3d5251 fix ordering in deduplication
82b00b848 fix garbage count in vivify for undeleting
7fdefad39 fixed infinite probe loop
4972c3a44 fix deduplicate sorting order
a8bf015bd correct statistic count of undeleting clauses
6a27f1098 add missing deallocation deduplicate_init
e3d5b1fc8 trigger deduplication for each solving
b9360c3aa fix examples with factor
4fd46f072 implement factorcheck option
afdd1c2bb fix printing
d83933c6d fix deleted constructor
c52686948 remove all garbage clauses from watch lists in deduplicate
37419d638 remove forwarded vacuous EquivalentTracer (#152)
24c6361ea add missing space in CFLAGS if user-defined (#146)
19f162b44 Make global variable atomic (#155)
e0598014c fixed typo in mobical parser
8b1631338 Fix: Explicitly use signed char in congruence.cpp for ARM compatibility
3e1cf598b avoid log10(0) (#150)
43fbdf760 fix scheduling
652b85c15 bump version
fa41dddd3 missing parsing of keyword
244239758 add optional deduplication of clauses initially
19051795f fixing memory leak in congruence
80fbeae73 fix memory leak in congruence
721932683 fix performance bug in fastelim due to wrong bounds

git-subtree-dir: cadical/cppsrc
git-subtree-split: 4198d817d0dcde5b1240eefbff70b555b7df2af9

+++ src/external.hpp
@@ -125,7 +125,7 @@ struct External {
   bool is_witness (int elit);
   bool is_decision (int elit);
 
-  void force_backtrack (size_t new_level);
+  void force_backtrack (int new_level);
 
   //----------------------------------------------------------------------//
